Captured EKF rmse result shown in simulator
Launched Udacity's Term2 simulator for the extended kalman project.
Executed ExtendedKF C++ program to run sensor fusion and calculate
rmse for sensors (radar, lidar). The simulator is the client and
the C++ program is a web server. Before I hit the start button in
the simulator, I selected Dataset 1, which starts with Lidar data,
then alternates to Radar and repeats until reaching end of data.
I hit the start button, allowed the simulator to run and the C++
program calculated rmse and simulator displayed it live on its GUI.
When the simulator finished and the car drove around the crazy 8
racetrack, the final rmse was output to the screen. I captured a
picture of it for Dataset 1. Then did the same process for Dataset
2. Screenshots show rmse (px,py,vx,py) meet Udacity project
specifications rmse <= [0.11, 0.11, 0.52, 0.52].